权限控制

+++ app/Service/Dao/DimensionDao.php
@@ -11,6 +11,8 @@ declare(strict_types=1);
  */
 namespace App\Service\Dao;
 
+use App\Constants\ErrorCode;
+use App\Exception\BusinessException;
 use App\Model\Dimension;
 use App\Service\Formatter\DimensionFormatter;
 use Hyperf\Di\Annotation\Inject;
@@ -27,16 +29,19 @@ class DimensionDao extends Dao
 
     public function createOrUpdate(int $userId, array $attributes, int $id = 0): Dimension
     {
-        $this->isLoader($userId);
+        $this->isLeader($userId);
         $model = $this->findById($id);
         if (empty($model)) {
             $model = new Dimension();
+            $model->leader_id = $userId;
+        }
+        if (! $model->own($userId)) {
+            throw new BusinessException(ErrorCode::OPERATION_INVALID);
         }
         $model->classify_id = $attributes['classify_id'];
         $model->review = $attributes['review'];
         $model->score = $attributes['score'];
         $model->review_description = $attributes['review_description'] ?? null;
-        $model->leader_id = $userId;
         $model->user_id = $attributes['user_id'];
         $model->save();
 
@@ -45,10 +50,13 @@ class DimensionDao extends Dao
 
     public function delete(int $userId, int $id): bool
     {
-        $this->isLoader($userId);
+        $this->isLeader($userId);
         $model = $this->findById($id);
         if (empty($model)) {
             return true;
+        }
+        if (! $model->own($userId)) {
+            throw new BusinessException(ErrorCode::OPERATION_INVALID);
         }
         $model->delete();
 
@@ -57,7 +65,7 @@ class DimensionDao extends Dao
 
     public function getWhereByAll(int $userId, ?string $review, int $offset = 0, int $limit = 10, array $columns = ['*']): array
     {
-        $this->isLoader($userId);
+        $this->isLeader($userId);
         $builder = Dimension::query();
         $builder->where('review', 'like', '%' . $review . '%')
             ->orderBy('id', 'desc');
